Add HTTP/2 signatures
Add a framework for building signatures for HTTP/2 clients. The
signature is based on the HTTP/2 frames the client sends at the
beginning of the session. The signature can currently be created from
the output of nghttpd (a simple HTTP/2 server).

Since there is common functionality to the HTTP/2 signatures and the
existing TLS signature framework, some of it was separated into a common
Signature class.